Split job file into multiple files for better overview
