+++ internal/model/job_amqp.go
@@ -1,0 +1,58 @@
+package model
+
+import error2 "github.com/xBlaz3kx/distributed-scheduler/internal/pkg/error"
+
+type BodyEncoding string
+
+const (
+	BodyEncodingBase64 BodyEncoding = "base64"
+)
+
+func (be *BodyEncoding) Valid() bool {
+	if be == nil {
+		return true
+	}
+	switch *be {
+	case BodyEncodingBase64:
+		return true
+	default:
+		return false
+	}
+}
+
+type AMQPJob struct {
+	// Todo encode the connection string!
+	Connection   string                 `json:"connection"`    // e.g., "amqp://guest:guest@localhost:5672/"
+	Exchange     string                 `json:"exchange"`      // e.g., "my_exchange"
+	RoutingKey   string                 `json:"routing_key"`   // e.g., "my_routing_key"
+	Headers      map[string]interface{} `json:"headers"`       // e.g., {"x-delay": 10000}
+	Body         string                 `json:"body"`          // e.g., "Hello, world!"
+	BodyEncoding *BodyEncoding          `json:"body_encoding"` // e.g., null, "base64"
+	ContentType  string                 `json:"content_type"`  // e.g., "text/plain"
+}
+
+// Validate validates an AMQPJob struct.
+func (amqpJob *AMQPJob) Validate() error {
+	if amqpJob == nil {
+		return error2.ErrAMQPJobNotDefined
+	}
+
+	// Todo validate URL
+	if amqpJob.Connection == "" {
+		return error2.ErrEmptyHTTPJobURL
+	}
+
+	if amqpJob.Exchange == "" {
+		return error2.ErrEmptyExchange
+	}
+
+	if amqpJob.RoutingKey == "" {
+		return error2.ErrEmptyRoutingKey
+	}
+
+	if !amqpJob.BodyEncoding.Valid() {
+		return error2.ErrInvalidBodyEncoding
+	}
+
+	return nil
+}
